pyproject.toml
[project]
name = "essay-writer"
version = "0.1.0"
description = "Deep agent system for writing academic essays in Greek"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.11",
    "jinja2>=3.1",
    "python-docx>=1.1",
    "pymupdf>=1.25",
    "python-pptx>=1.0.2",
    "pydantic-settings[yaml]>=2.13.1",
]
